streamlit ui code and batch processing

+++ src/utils.py
@@ -5,7 +5,7 @@ import pandas as pd
 # File Discovery
 # ============================================================
 
-def discover_files(input_dir: Path, patterns: List = ["*.pdf", "*.txt","*.docx"]) -> pd.DataFrame:
+def discover_files(input_dir: Path, patterns: List = ["*.pdf", "*.txt","*.docx"], number_of_files=None) -> pd.DataFrame:
     """
     Discovers files inside a provided path and returns
     a standardized DataFrame.
@@ -17,12 +17,18 @@ def discover_files(input_dir: Path, patterns: List = ["*.pdf", "*.txt","*.docx"]
     Returns:
         pd.DataFrame
     """
-
-    files = sorted(
-        file
-        for pattern in patterns
-        for file in input_dir.glob(pattern)
-    )
+    if number_of_files:
+        files = sorted(
+            file
+            for pattern in patterns
+            for file in input_dir.glob(pattern)
+        )[:number_of_files]
+    else:
+        files = sorted(
+            file
+            for pattern in patterns
+            for file in input_dir.glob(pattern)
+        )
 
     if not files:
         raise FileNotFoundError(
@@ -37,7 +43,7 @@ def discover_files(input_dir: Path, patterns: List = ["*.pdf", "*.txt","*.docx"]
                 round(file.stat().st_size / 1024, 2)
                 for file in files
             ],
-            "status": "Pending",
+            "parsed_status": "Pending",
         }
     )
 
@@ -54,12 +60,31 @@ def load_prompt(prompt_path: Path) -> str:
     with open(prompt_path, "r", encoding="utf-8") as file:
         return file.read()
 
+# def save_dataframe(
+#     dataframe: pd.DataFrame,
+#     output_path: Path,
+# ):
+#     """
+#     Saves a DataFrame to CSV.
+#     """
+
+#     output_path.parent.mkdir(
+#         parents=True,
+#         exist_ok=True,
+#     )
+
+#     dataframe.to_csv(
+#         output_path,
+#         index=False,
+#     )
+
 def save_dataframe(
     dataframe: pd.DataFrame,
     output_path: Path,
 ):
     """
     Saves a DataFrame to CSV.
+    Appends to the existing file if it already exists.
     """
 
     output_path.parent.mkdir(
@@ -67,7 +92,18 @@ def save_dataframe(
         exist_ok=True,
     )
 
-    dataframe.to_csv(
-        output_path,
-        index=False,
-    )
+    # Append if file already exists
+    if output_path.exists():
+        dataframe.to_csv(
+            output_path,
+            mode="a",
+            header=False,
+            index=False,
+        )
+    else:
+        dataframe.to_csv(
+            output_path,
+            mode="w",
+            header=True,
+            index=False,
+        )
